Add passive auth support for `login()`
This depends on changes in pysaml2 as well, but this exposes adding
IsPassive=true to the SAML request if you pass in `is_passive='true'` to
login, where it will redirect back silently if it cannot automatically
authenticate the user.

This also adds a `redirect_to` argument to control returning to a
speccific, perhaps current, location.